Moves Ground source and header files to /src and /include
